move class into data layer
